requirements.txt
# VR-Bench Dependencies
# Python >= 3.10

# Core dependencies
numpy>=1.24.0
opencv-python>=4.8.0
pillow>=10.0.0
pyyaml>=6.0
imageio>=2.31.0
imageio-ffmpeg>=0.4.9

# Data processing
scipy>=1.11.0
scikit-image>=0.21.0

# VLM evaluation
openai>=1.0.0
python-dotenv>=1.0.0

# Optional: GPU acceleration
# cupy-cuda12x>=12.0.0  # For CUDA 12.x
# cupy-cuda11x>=11.0.0  # For CUDA 11.x

# Development dependencies (optional)
# pytest>=7.4.0
# black>=23.0.0
# flake8>=6.0.0
# mypy>=1.5.0
# pylint>=2.17.0

